:recycle: Rename to pill sheet type column
